Update style.css and views
This commit updates the style.css file to change the height property of the .m-v-v class to max-content. It also removes the Apply online button from the admission.hbs view and updates the contact information in the offices.hbs view. Additionally, it updates the text in the user-homePage.hbs view to better describe the values of Darul Irfan. Finally, it updates the link in the user-header.hbs partial to point to the maintenance page instead of the offices page.